More methods added to the Elastic wrapper (in progress)
